@@ -2,32 +2,34 @@
   "name": "elysia-docs",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "main": "index.js",
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@iconify/json": "^2.2.63",
+    "@iconify/json": "^2.2.104",
     "@tailwindcss/postcss7-compat": "^2.2.17",
-    "autoprefixer": "^10.4.14",
-    "openai": "^3.2.1",
-    "postcss": "^8.4.23",
+    "autoprefixer": "^10.4.15",
+    "daisyui": "^3.5.1",
+    "elysia": "^0.6.11",
+    "openai": "^4.0.1",
+    "postcss": "^8.4.28",
     "prismjs": "^1.29.0",
-    "tailwindcss": "^3.3.2",
-    "vite": "^4.3.5",
+    "tailwindcss": "^3.3.3",
+    "vite": "^4.4.9",
     "vite-plugin-icons": "^0.6.5",
-    "vitepress-plugin-search": "^1.0.4-alpha.20",
     "vue-prism-component": "^2.0.0",
     "windicss": "^3.5.6"
   },
   "devDependencies": {
-    "@types/node": "^20.1.2",
-    "bun-types": "^0.5.8",
+    "@types/node": "^20.5.1",
+    "bun-types": "^0.7.3",
     "cheerio": "1.0.0-rc.12",
     "flexsearch": "^0.7.31",
     "markdown-it": "^13.0.1",
-    "vitepress": "1.0.0-alpha.75",
-    "vue": "^3.2.47"
+    "vitepress": "1.0.0-rc.4",
+    "vue": "^3.3.4"
   },
   "scripts": {
     "dev": "vitepress dev docs",
